+++ tests/core/test_lightning_optimizer.py
@@ -20,7 +20,7 @@ from torch.optim import Adam, Optimizer, SGD
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.core.optimizer import LightningOptimizer
-from pytorch_lightning.loops.closure import LightningClosure
+from pytorch_lightning.loops.closure import Closure
 from tests.helpers.boring_model import BoringModel
 
 
@@ -222,7 +222,7 @@ def test_lightning_optimizer_automatic_optimization_optimizer_step(tmpdir):
             ...
 
         def optimizer_step(self, epoch, batch_idx, optimizer, optimizer_idx, optimizer_closure, **_):
-            assert isinstance(optimizer_closure, LightningClosure)
+            assert isinstance(optimizer_closure, Closure)
             # not passing the closure to the optimizer because step is mocked
             # zero_grad is called inside the closure
             if isinstance(optimizer, SGD) and batch_idx % 2 == 0: